@@ -1,6 +1,6 @@
 {
   "name": "apigee-x-module",
-  "version": "0.7.4",
+  "version": "1.0.0",
   "description": "Module for integrating with the Apigee X platform for TS/JS developer portal integrations.",
   "homepage": "https://github.com/tyayers/apigee-x-module",
   "bugs": "https://github.com/tyayers/apigee-x-module/issues",
@@ -11,29 +11,33 @@
     "gcp",
     "api management"
   ],
-  "main": "dist/src/index.js",
-  "types": "dist/src/index.d.ts",
+  "type": "module",
+  "exports": "./dist/src/index.js",
+  "types": "./dist/src/index.d.ts",
+  "engines": {
+    "node": ">=16"
+  },
   "files": [
     "/dist"
   ],
   "scripts": {
     "build": "rm -rf dist/** && tsc",
-    "test": "mocha -r ts-node/register test/**/*.test.ts --timeout 5000",
+    "test": "mocha",
     "deploy": "npm run build && npm publish"
   },
   "dependencies": {
-    "axios": "^0.24.0",
+    "axios": "^1.6.2",
     "dotenv": "^10.0.0",
     "form-data": "^4.0.0",
-    "google-auth-library": "^7.9.2"
+    "google-auth-library": "^9.2.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.2.22",
-    "@types/mocha": "^9.0.0",
-    "@types/node": "^16.11.32",
-    "chai": "^4.3.4",
-    "mocha": "^9.1.1",
-    "ts-node": "^10.2.1",
-    "typescript": "^4.4.3"
+    "@types/chai": "^4.3.10",
+    "@types/mocha": "^10.0.4",
+    "@types/node": "^20.9.2",
+    "chai": "^4.3.10",
+    "mocha": "10.2.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.2.2"
   }
 }
